backend/requirements.txt
# Backend dependencies for Catan Game API
# 
# IMPORTANT: Always use 'uv pip' (not 'pip') to install packages
# Example: source .venv/bin/activate && uv pip install -r requirements.txt
# See backend/DEVELOPMENT.md for more details

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.0
litellm>=1.0.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

